Some minor cleanup in the operator
- Moved the cluster upgrade test from test_operations.py to test_upgrade.py
- Refactored some tests that were provisioning ad-hoc CrateDBs to use the utility method that does it for you
- Removed the unused cleanup_handler parameter from start_cluster() (was a remnant of crate-local)
